refactor(cypress): expose oidc helper from appshell/ssr bundle
