+++ packages/application-shell/src/components/authenticated/has-cached-authentication-state.ts
@@ -1,8 +1,8 @@
 import type { ApplicationWindow } from '@commercetools-frontend/constants';
 
 import { STORAGE_KEYS, OIDC_CLAIMS } from '../../constants';
+import { buildOidcScope } from '../../utils/oidc';
 import * as oidcStorage from '../../utils/oidc-storage';
-import { buildOidcScope } from './helpers';
 
 declare let window: ApplicationWindow;
