+++ packages/application-shell/src/ssr/index.ts
@@ -1,1 +1,2 @@
 export * from '../utils/formatters';
+export { buildOidcScope } from '../utils/oidc';
