+++ packages/cypress/package.json
@@ -38,6 +38,7 @@
     "@babel/runtime": "7.16.5",
     "@babel/runtime-corejs3": "7.16.5",
     "@commercetools-frontend/application-config": "20.12.1",
+    "@commercetools-frontend/application-shell": "20.12.2",
     "@manypkg/get-packages": "1.1.3",
     "uuid": "8.3.2"
   },
