+++ packages/cypress/src/constants.ts
@@ -5,10 +5,3 @@ export const STORAGE_KEYS = {
 } as const;
 
 export const OIDC_RESPONSE_TYPES = { ID_TOKEN: 'id_token' };
-export const OIDC_CLAIMS = {
-  OPEN_ID: 'openid',
-  PROJECT_KEY: 'project_key',
-  TEAM_ID: 'team_id',
-  VIEW: 'view',
-  MANAGE: 'manage',
-};
